+++ app/controllers/external_controller.rb
@@ -29,6 +29,16 @@ class ExternalController < ApplicationController
     end
   end
 
+  # GET /sitemap.xml.gz
+  def sitemap_xml
+    sitemap_xml = File.join(CarrierWave::Uploader::Base.root, 'sitemaps', 'sitemap.xml.gz')
+    if File.exist?(sitemap_xml)
+      send_file sitemap_xml
+    else
+      head :ok
+    end
+  end
+
   private
 
   def set_design